+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog 6.1.0
 - Environments use Slot Modifications instead of merge cards
+- Reworked Enviroment handling and updated some Enviroments
 - SFX now uses API
 - Dialogue now uses API